Extract Item
